refactor: model 파일명 store로 수정
- state 관리에 적합한 이름으로 수정
- 역할을 명확하게 하기 위해 디렉토리 분리
